Fix timestamp error on storage syncing
