+++ app/src/main/java/org/qosp/notes/ui/MainActivity.kt
@@ -19,6 +19,7 @@ import androidx.navigation.NavDeepLinkBuilder
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.navigateUp
+import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.runBlocking
@@ -26,6 +27,7 @@ import org.koin.androidx.viewmodel.ext.android.viewModel
 import org.qosp.notes.R
 import org.qosp.notes.components.backup.BackupService
 import org.qosp.notes.data.model.Notebook
+import org.qosp.notes.data.sync.nextcloud.NextcloudConfig
 import org.qosp.notes.databinding.ActivityMainBinding
 import org.qosp.notes.preferences.SortNavdrawerNotebooksMethod
 import org.qosp.notes.ui.utils.closeAndThen
@@ -152,14 +154,12 @@ class MainActivity : BaseActivity() {
             }
         }
 
-        // TODO: Check if this is really being used or can be removed
-//        syncManager.config
-//            .collect(this@MainActivity) { config ->
-//                if (config is NextcloudConfig) {
-//                    textViewUsername.text = config.username
-//                    textViewProvider.text = getString(config.provider.nameResource)
-//                }
-//            }
+        NextcloudConfig.fromPreferences(preferenceRepository)
+            .filterNotNull()
+            .collect(this@MainActivity) { config ->
+                textViewUsername.text = config.username
+                textViewProvider.text = getString(R.string.preferences_cloud_service_nextcloud)
+            }
     }
 
     private fun selectCurrentDestinationMenuItem(
